feat/메뉴 삭제 API 구현

+++ src/main/java/com/itwill/matzip/dto/MenusToCreate.java
@@ -12,6 +12,6 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class MenusToCreate {
-    Integer restaurantId;
+    Long restaurantId;
     List<MenuToCreate> menus;
 }
